Here I deleted the java_lang_Math folder and the math_api folder and replaced them with the Math folder.
