+++ Math/Logarithm.java
