--- java_lang_Math/exponentiation/Cbrt.java
@@ -1,11 +1,0 @@
-public class Cbrt
-{
-  /**
-   * @requires Math.isFinite(num);
-   * @ensures \result = Math.cbrt(num)
-   */
-  public static double cbrtDouble(double num)
-  {
-    return Math.cbrt(num);
-  }
-}
