--- java_lang_Math/exponentiation/Exp.java
@@ -1,10 +1,0 @@
-public class Exp
-{
-  /**
-   * @ensures \result = Math.exp(num)
-   */
-  public static double expDouble(double num)
-  {
-    return Math.exp(num);
-  }
-}
