--- java_lang_Math/exponentiation/Expml.java
@@ -1,10 +1,0 @@
-public class Expml
-{
-  /**
-   * @ensures \result = Math.expml(num)
-   */
-  public static double expm1Double(double num)
-  {
-    return Math.expm1(num);
-  }
-}
