--- java_lang_Math/exponentiation/GetExponent.java
@@ -1,10 +1,0 @@
-public class GetExponent
-{
-  /**
-   * @ensures \result = Math.getExponent(num);
-   */
-  public static double getExponenetDouble(double num)
-  {
-    return Math.getExponent(num);
-  }
-}
