--- java_lang_Math/exponentiation/Pow.java
@@ -1,10 +1,0 @@
-public class Pow
-{
-  /**
-   * @eunsures \result = Math.pow(num, pow);
-   */
-  public static double powDouble(double num, doulbe pow)
-  {
-    return Math.pow(num, pow);
-  }
-}
