--- java_lang_Math/exponentiation/Sqrt.java
@@ -1,11 +1,0 @@
-public class Sqrt
-{
-  /**
-   * @requires (Math.isFinite(num));
-   * @ensures \result = Math.sqrt(num);
-   */
-  public static double sqrtDouble(double num)
-  {
-    return Math.sqrt(num);
-  }
-}
