--- java_lang_Math/misc/Max.output
@@ -1,58 +1,0 @@
-Proving methods in Max
-Starting proof of Max.Max() with prover z3_4_3
-Method assertions are validated [1.06 secs]
-Max.java:3: Feasibility check #1 - end of preconditions : OK [0.00 secs]
-Max.java:3: Feasibility check #2 - at program exit : OK [0.00 secs]
-Completed proof of Max.Max() with prover z3_4_3 - no warnings [1.14 secs]
-Starting proof of Max.max_test_1(double,double) with prover z3_4_3
-Max.max_test_1 Method assertions are INVALID
-Max.java:17: warning: The prover cannot establish an assertion (Postcondition: Max.java:10: ) in method max_test_1
-	return c;
-	^
-Max.java:10: warning: Associated declaration: Max.java:17: 
-//@ ensures \result == b;
-    ^
-Max.java:17: warning: The prover cannot establish an assertion (Postcondition: Max.java:13: ) in method max_test_1
-	return c;
-	^
-Max.java:13: warning: Associated declaration: Max.java:17: 
-//@ ensures \result == a;
-    ^
-Completed proof of Max.max_test_1(double,double) with prover z3_4_3 - with warnings [1.89 secs]
-Starting proof of Max.max_test_2(double,double) with prover z3_4_3
-Max.max_test_2 Method assertions are INVALID
-Max.java:25: warning: The prover cannot establish an assertion (Postcondition: Max.java:20: ) in method max_test_2
-        return c;
-        ^
-Max.java:20: warning: Associated declaration: Max.java:25: 
-//@ensures \result >= a && \result >= b;
-   ^
-Completed proof of Max.max_test_2(double,double) with prover z3_4_3 - with warnings [1.67 secs]
-Starting proof of Max.max_test_3(double,double) with prover z3_4_3
-Max.max_test_3 Method assertions are INVALID
-Max.java:33: warning: The prover cannot establish an assertion (Postcondition: Max.java:28: ) in method max_test_3
-	return c;
-	^
-Max.java:28: warning: Associated declaration: Max.java:33: 
-//@ensures \result >= 5*a - 4 && \result >= -2*b + 3;
-   ^
-Completed proof of Max.max_test_3(double,double) with prover z3_4_3 - with warnings [1.63 secs]
-Starting proof of Max.max_test_4(double,double) with prover z3_4_3
-Method assertions are validated [1.39 secs]
-Max.java:36: Feasibility check #1 - end of preconditions : OK [0.11 secs]
-Max.java:36: Feasibility check #2 - at program exit : OK [0.08 secs]
-Completed proof of Max.max_test_4(double,double) with prover z3_4_3 - no warnings [1.66 secs]
-Completed proving methods in Max [8.04 secs]
-Note: Summary:
-    Valid:        2
-    Invalid:      3
-    Infeasible:   0
-    Timeout:      0
-    Error:        0
-    Skipped:      0
-   TOTAL METHODS: 5
-   Classes:       0 proved of 1
-   Model Classes: 0
-   Model methods: 0 proved of 0
-   DURATION:          8.6 secs
-8 warnings
