--- java_lang_Math/misc/Min.output
@@ -1,64 +1,0 @@
-Proving methods in Min
-Starting proof of Min.Min() with prover z3_4_3
-Method assertions are validated [1.06 secs]
-Min.java:3: Feasibility check #1 - end of preconditions : OK [0.00 secs]
-Min.java:3: Feasibility check #2 - at program exit : OK [0.00 secs]
-Completed proof of Min.Min() with prover z3_4_3 - no warnings [1.13 secs]
-Starting proof of Min.min_test_1(double,double) with prover z3_4_3
-Min.min_test_1 Method assertions are INVALID
-Min.java:17: warning: The prover cannot establish an assertion (Postcondition: Min.java:10: ) in method min_test_1
-	return c;
-	^
-Min.java:10: warning: Associated declaration: Min.java:17: 
-//@ ensures \result == a;
-    ^
-Min.java:17: warning: The prover cannot establish an assertion (Postcondition: Min.java:13: ) in method min_test_1
-	return c;
-	^
-Min.java:13: warning: Associated declaration: Min.java:17: 
-//@ ensures \result == b;
-    ^
-Completed proof of Min.min_test_1(double,double) with prover z3_4_3 - with warnings [1.84 secs]
-Starting proof of Min.min_test_2(double,double) with prover z3_4_3
-Min.min_test_2 Method assertions are INVALID
-Min.java:26: warning: The prover cannot establish an assertion (Postcondition: Min.java:20: ) in method min_test_2
-        return c;
-        ^
-Min.java:20: warning: Associated declaration: Min.java:26: 
-//@ ensures \result <= a;
-    ^
-Min.java:26: warning: The prover cannot establish an assertion (Postcondition: Min.java:21: ) in method min_test_2
-        return c;
-        ^
-Min.java:21: warning: Associated declaration: Min.java:26: 
-//@ ensures \result <= b;
-    ^
-Completed proof of Min.min_test_2(double,double) with prover z3_4_3 - with warnings [1.82 secs]
-Starting proof of Min.min_test_3(double,double) with prover z3_4_3
-Min.min_test_3 Method assertions are INVALID
-Min.java:34: warning: The prover cannot establish an assertion (Postcondition: Min.java:29: ) in method min_test_3
-	return c;
-	^
-Min.java:29: warning: Associated declaration: Min.java:34: 
-//@ ensures \result <= 5*a - 4 && \result <= -2*b + 3;
-    ^
-Completed proof of Min.min_test_3(double,double) with prover z3_4_3 - with warnings [1.57 secs]
-Starting proof of Min.min_test_4(double,double) with prover z3_4_3
-Method assertions are validated [1.35 secs]
-Min.java:37: Feasibility check #1 - end of preconditions : OK [0.14 secs]
-Min.java:37: Feasibility check #2 - at program exit : OK [0.13 secs]
-Completed proof of Min.min_test_4(double,double) with prover z3_4_3 - no warnings [1.68 secs]
-Completed proving methods in Min [8.09 secs]
-Note: Summary:
-    Valid:        2
-    Invalid:      3
-    Infeasible:   0
-    Timeout:      0
-    Error:        0
-    Skipped:      0
-   TOTAL METHODS: 5
-   Classes:       0 proved of 1
-   Model Classes: 0
-   Model methods: 0 proved of 0
-   DURATION:          8.6 secs
-10 warnings
