--- java_lang_Math/misc/Random.output
@@ -1,48 +1,0 @@
-Proving methods in Random
-Starting proof of Random.Random() with prover z3_4_3
-Method assertions are validated [1.07 secs]
-Random.java:2: Feasibility check #1 - end of preconditions : OK [0.00 secs]
-Random.java:2: Feasibility check #2 - at program exit : OK [0.00 secs]
-Completed proof of Random.Random() with prover z3_4_3 - no warnings [1.14 secs]
-Starting proof of Random.random_test_1() with prover z3_4_3
-Method assertions are validated [1.04 secs]
-Random.java:10: Feasibility check #1 - end of preconditions : OK [0.00 secs]
-Random.java:10: Feasibility check #2 - at program exit : OK [0.00 secs]
-Completed proof of Random.random_test_1() with prover z3_4_3 - no warnings [1.08 secs]
-Starting proof of Random.random_test_2() with prover z3_4_3
-Method assertions are validated [1.02 secs]
-Random.java:18: Feasibility check #1 - end of preconditions : OK [0.00 secs]
-Random.java:18: Feasibility check #2 - at program exit : OK [0.00 secs]
-Completed proof of Random.random_test_2() with prover z3_4_3 - no warnings [1.05 secs]
-Starting proof of Random.random_test_3() with prover z3_4_3
-Random.random_test_3 Method assertions are INVALID
-Random.java:32: warning: The prover cannot establish an assertion (Postcondition: Random.java:28: ) in method random_test_3
-                return a;
-                ^
-Random.java:28: warning: Associated declaration: Random.java:32: 
-	//@ensures 0.0 > \result || \result >= 1.0;
-	   ^
-Completed proof of Random.random_test_3() with prover z3_4_3 - with warnings [1.07 secs]
-Starting proof of Random.random_test_4() with prover z3_4_3
-Random.random_test_4 Method assertions are INVALID
-Random.java:42: warning: The prover cannot establish an assertion (Postcondition: Random.java:36: ) in method random_test_4
-                return a;
-                ^
-Random.java:36: warning: Associated declaration: Random.java:42: 
-	//@ensures -5.0 > \result || \result >= 5.0;
-	   ^
-Completed proof of Random.random_test_4() with prover z3_4_3 - with warnings [1.04 secs]
-Completed proving methods in Random [5.44 secs]
-Note: Summary:
-    Valid:        3
-    Invalid:      2
-    Infeasible:   0
-    Timeout:      0
-    Error:        0
-    Skipped:      0
-   TOTAL METHODS: 5
-   Classes:       0 proved of 1
-   Model Classes: 0
-   Model methods: 0 proved of 0
-   DURATION:          5.6 secs
-4 warnings
