--- java_lang_Math/trig/Sine.java
@@ -1,14 +1,0 @@
-// written by Robert Moore 5/23/21
-
-public class Sine {
-  //@ requires Math.isFinite(a);
-  //@ ensures \result >= -1.0 && \result <= 1.0;
-  static double inBounds(double a){
-    return Math.sin(a);
-  }
-
-  //@ ensures \result == 0.0;
-  static double sinZero(){
-    return Math.sin(0.0);
-  }
-}
